pyproject.toml
[project]
name = "budget-tool"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "enablebanking_sdk>=0.1.6",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
